Added two Charts: Faceoff and GoalsByPeriod
Added two bar charts for : Faceoff Win % Breakdown and Goals By Period (For and Against). Also added some color change by making a function that gets the opposite color of the team color and put it the graph. Thinking of just calling the function on the team page and then sending it to the component instead.