Calendar of Games
Added Component for viewing the calendar of games for selected team. For upcoming games the calendar will show "vs or @" depending on AWAY or HOME for the selected team. It will also show the start time of the game. For completed games in the past it will show the score and if the game has not ended in regulation it will show either OT or SO for overtime and shootout respectively.