Player Page plus Components
Worked on the Player Page making it functional to click on a player in the teams Page and the landing on the specific Player Page with basic player info. Made two components one for featured stats (cards) and the other for season totals (table). Also added API call in backend for the stats currently used for the player page.